Load profile description from JSON and modularize assets
