+++ index.html
@@ -22,9 +22,7 @@
                 <!-- Profile Section -->
                 <section class="mb-6">
                     <h3 class="text-2xl font-bold mb-4 text-edge-green flex items-center"><i class="fas fa-user-tie mr-3 text-edge-cyan"></i>PERFIL PROFESIONAL</h3>
-                    <p class="text-gray-600 leading-relaxed text-sm">
-                        Consultor y gestor administrativo con sólida trayectoria en los sectores financiero, comercial y digital. Con amplias competencias en inteligencia de negocios, atención al cliente, procesos documentales y gestión de conflictos. Cuento, además, con una base formativa en Derecho colombiano gracias a seis semestres cursados, lo cual refuerza mi criterio en la interpretación normativa y en la gestión de procesos públicos. Destaco por mi habilidad para generar soluciones eficaces, adaptarme a contextos complejos y fomentar relaciones humanas basadas en la empatía, el servicio y la comunicación asertiva. Mi experiencia incluye liderar procesos de expansión, atención ciudadana, resolución de trámites, formación educativa y manejo de herramientas tecnológicas para optimizar la operación institucional. Mi formación académica, en combinación con mi vocación de servicio, me perfilan como un aliado estratégico para dependencias públicas de carácter administrativo, social o financiero.
-                    </p>
+                    <p id="profileDescription" class="text-gray-600 leading-relaxed text-sm"></p>
                 </section>
 
                 <!-- Experience Section -->
@@ -66,6 +64,6 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/html2canvas/1.4.1/html2canvas.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/2.5.1/jspdf.umd.min.js"></script>
 
-    <script src="scripts/main.js"></script>
+    <script type="module" src="scripts/index.js"></script>
     </body>
 </html>
