+++ profile.json
@@ -1,6 +1,7 @@
 {
   "name": "Rayo Palatianos",
   "title": "Consultor y Gestor Administrativo",
+  "description": "Consultor y gestor administrativo con sólida trayectoria en los sectores financiero, comercial y digital. Con amplias competencias en inteligencia de negocios, atención al cliente, procesos documentales y gestión de conflictos. Cuento, además, con una base formativa en Derecho colombiano gracias a seis semestres cursados, lo cual refuerza mi criterio en la interpretación normativa y en la gestión de procesos públicos. Destaco por mi habilidad para generar soluciones eficaces, adaptarme a contextos complejos y fomentar relaciones humanas basadas en la empatía, el servicio y la comunicación asertiva. Mi experiencia incluye liderar procesos de expansión, atención ciudadana, resolución de trámites, formación educativa y manejo de herramientas tecnológicas para optimizar la operación institucional. Mi formación académica, en combinación con mi vocación de servicio, me perfilan como un aliado estratégico para dependencias públicas de carácter administrativo, social o financiero.",
   "photoUrl": "https://media.licdn.com/dms/image/v2/D4D03AQFs-EKWXsDmMg/profile-displayphoto-shrink_100_100/profile-displayphoto-shrink_100_100/0/1709139849249?e=2147483647&v=beta&t=N65UtqTk2eSflBRG_EFU9-wtfhW-5duQVjSaxW_ApoQ",
   "contact": {
     "email": "rayo.palatianos@gmail.com",
