+++ scripts/index.js
@@ -1,0 +1,7 @@
+import { setupPdfExport } from './pdf.js';
+import { loadData } from './dataLoader.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+  setupPdfExport();
+  loadData();
+});
