+++ styles/base.css
@@ -1,0 +1,13 @@
+body {
+    font-family: 'Inter', sans-serif;
+    background-color: #f0f4f8;
+}
+.edge-blue { background-color: #0078D4; }
+.bg-adobe-red {
+    background-color: #ff0000;
+    background-image: linear-gradient(45deg, #ff4b4b, #ff0000, #b20000);
+}
+.text-edge-green { color: #42C89A; }
+.text-edge-cyan { color: #00A4EF; }
+.border-edge-green { border-color: #42C89A; }
+.icon-cyan { color: #00A4EF; }
