+++ styles/button.css
@@ -1,0 +1,20 @@
+#exportPdfBtn {
+    position: relative;
+    overflow: hidden;
+}
+#exportPdfBtn::after {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: -100%;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(120deg, transparent, rgba(255,255,255,0.4), transparent);
+    pointer-events: none;
+    animation: shine 3s infinite;
+}
+@keyframes shine {
+    0% { left: -100%; }
+    60% { left: 100%; }
+    100% { left: 100%; }
+}
