+++ styles/main.css
@@ -1,62 +1,3 @@
-body {
-    font-family: 'Inter', sans-serif;
-    background-color: #f0f4f8;
-}
-.edge-blue { background-color: #0078D4; }
-.bg-adobe-red {
-    background-color: #ff0000;
-    background-image: linear-gradient(45deg, #ff4b4b, #ff0000, #b20000);
-}
-.text-edge-green { color: #42C89A; }
-.text-edge-cyan { color: #00A4EF; }
-.border-edge-green { border-color: #42C89A; }
-.icon-cyan { color: #00A4EF; }
-#exportPdfBtn {
-    position: relative;
-    overflow: hidden;
-}
-#exportPdfBtn::after {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: -100%;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(120deg, transparent, rgba(255,255,255,0.4), transparent);
-    pointer-events: none;
-    animation: shine 3s infinite;
-}
-@keyframes shine {
-    0% { left: -100%; }
-    60% { left: 100%; }
-    100% { left: 100%; }
-}
-.timeline-item {
-    position: relative;
-    padding-bottom: 2rem;
-    padding-left: 2.5rem;
-}
-.timeline-item:last-child {
-    padding-bottom: 0;
-}
-.timeline-item::before {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: 5px;
-    height: 100%;
-    width: 2px;
-    background-color: #e2e8f0;
-}
-.timeline-item::after {
-    content: '';
-    position: absolute;
-    left: -6px;
-    top: 5px;
-    height: 14px;
-    width: 14px;
-    border-radius: 50%;
-    background-color: #00A4EF;
-    border: 2px solid #fff;
-    z-index: 10;
-}
+@import url('base.css');
+@import url('button.css');
+@import url('timeline.css');
